requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=8.4.0
tqdm>=4.65.0
tabulate>=0.9.0
psutil>=5.8.0
kaggle>=1.5.13
requests>=2.28.0
python-dateutil>=2.8.2